rag/promptflow/requirements.txt
promptflow
promptflow-tools
python-dotenv
faiss-cpu
PyPDF2